backend/python-service/requirements.txt
fastapi
uvicorn
python-dotenv
requests
scikit-learn
joblib
